Added ops-inject and network fuzzing sota content
Reworked chapter 3 in order to add the cscs paper as well.

Still have to fill in the problem statement of each paper section, as
well as write an introduction for the chapter as a whole, and link the
two together.

For the ops-inject paper, I added the related work subsection since
there isn't one for the sota paper. So no need to merge them. Instead,
make sure to redo the figures in draw.io and give them some color. Also,
check if the font size can be modified in the pgf file.